Move Data, PlaceRef into JArrow
